Restructure tools to allow dynamic (plugable) loading
