+++ tools.py
@@ -1,8 +1,11 @@
 #!/usr/bin/env python3
 """ The master tools.py script """
+import os
 import sys
+
+from importlib import import_module
+
 # Importing the various tools
-import tools.cli as cli
 from lib.cli import FullHelpArgumentParser
 
 # Python version check
@@ -17,6 +20,18 @@ def bad_args(args):  # pylint:disable=unused-argument
     PARSER.print_help()
     exit(0)
 
+def _get_cli_opts():
+    """ Optain the subparsers and cli options for available tools """
+    base_path = os.path.realpath(os.path.dirname(sys.argv[0]))
+    tools_dir = os.path.join(base_path, "tools")
+    for tool_name in sorted(os.listdir(tools_dir)):
+        cli_file = os.path.join(tools_dir, tool_name, "cli.py")
+        if os.path.exists(cli_file):
+            mod = ".".join(("tools", tool_name, "cli"))
+            module = import_module(mod)
+            cliarg_class = getattr(module, "CliArgs")
+            help_text = getattr(module, "_HELPTEXT")
+            yield tool_name, help_text, cliarg_class
 
 if __name__ == "__main__":
     _TOOLS_WARNING = "Please backup your data and/or test the tool you want "
@@ -26,26 +41,8 @@ if __name__ == "__main__":
 
     PARSER = FullHelpArgumentParser()
     SUBPARSER = PARSER.add_subparsers()
-    ALIGN = cli.AlignmentsArgs(SUBPARSER,
-                               "alignments",
-                               "This command lets you perform various tasks pertaining to an "
-                               "alignments file.")
-    PREVIEW = cli.PreviewArgs(SUBPARSER,
-                              "preview",
-                              "This command allows you to preview swaps to tweak convert "
-                              "settings.")
-    EFFMPEG = cli.EffmpegArgs(SUBPARSER,
-                              "effmpeg",
-                              "This command allows you to easily execute common ffmpeg tasks.")
-    MASK = cli.MaskArgs(SUBPARSER,
-                        "mask",
-                        "This command lets you generate masks for existing alignments.")
-    RESTORE = cli.RestoreArgs(SUBPARSER,
-                              "restore",
-                              "This command lets you restore models from backup.")
-    SORT = cli.SortArgs(SUBPARSER,
-                        "sort",
-                        "This command lets you sort images using various methods.")
+    for tool, helptext, cli_args in _get_cli_opts():
+        cli_args(SUBPARSER, tool, helptext)
     PARSER.set_defaults(func=bad_args)
     ARGUMENTS = PARSER.parse_args()
     ARGUMENTS.func(ARGUMENTS)
