+++ tools/alignments/alignments.py
@@ -1,9 +1,12 @@
 #!/usr/bin/env python3
 """ Tools for manipulating the alignments seralized file """
+import sys
 import logging
-from .lib_alignments import (AlignmentData, Check, Dfl, Draw, # noqa pylint: disable=unused-import
-                             Extract, Fix, Manual, Merge, Rename,
-                             RemoveAlignments, Sort, Spatial, UpdateHashes)
+
+from .media import AlignmentData
+from .jobs import (Check, Dfl, Draw, Extract, Fix, Merge,  # noqa pylint: disable=unused-import
+                   Rename, RemoveAlignments, Sort, Spatial, UpdateHashes)
+from .jobs_manual import Manual # noqa pylint: disable=unused-import
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -21,10 +24,10 @@ class Alignments():
         logger.debug("Loading alignments")
         if len(self.args.alignments_file) > 1 and self.args.job != "merge":
             logger.error("Multiple alignments files are only permitted for merging")
-            exit(0)
+            sys.exit(0)
         if len(self.args.alignments_file) == 1 and self.args.job == "merge":
             logger.error("More than one alignments file required for merging")
-            exit(0)
+            sys.exit(0)
 
         if len(self.args.alignments_file) == 1:
             retval = AlignmentData(self.args.alignments_file[0])
