--- tools/lib_alignments/__init__.py
@@ -1,4 +1,0 @@
-from tools.lib_alignments.media import AlignmentData, ExtractedFaces, Faces, Frames
-from tools.lib_alignments.annotate import Annotate
-from tools.lib_alignments.jobs import Check, Dfl, Draw, Extract, Fix, Merge, RemoveAlignments, Rename, Sort, Spatial, UpdateHashes
-from tools.lib_alignments.jobs_manual import Manual
