+++ tools/restore/restore.py
@@ -3,6 +3,7 @@
 
 import logging
 import os
+import sys
 
 from lib.model.backup_restore import Backup
 
@@ -29,19 +30,19 @@ class Restore():
         """ Make sure there is only one model in the target folder """
         if not os.path.exists(self.model_dir):
             logger.error("Folder does not exist: '%s'", self.model_dir)
-            exit(1)
+            sys.exit(1)
         chkfiles = [fname for fname in os.listdir(self.model_dir) if fname.endswith("_state.json")]
         bkfiles = [fname for fname in os.listdir(self.model_dir) if fname.endswith(".bk")]
         if not chkfiles:
             logger.error("Could not find a model in the supplied folder: '%s'", self.model_dir)
-            exit(1)
+            sys.exit(1)
         if len(chkfiles) > 1:
             logger.error("More than one model found in the supplied folder: '%s'", self.model_dir)
-            exit(1)
+            sys.exit(1)
         if not bkfiles:
             logger.error("Could not find any backup files in the supplied folder: '%s'",
                          self.model_dir)
-            exit(1)
+            sys.exit(1)
         self.model_name = chkfiles[0].replace("_state.json", "")
         logger.info("%s Model found", self.model_name.title())
         logger.verbose("Backup files: %s)", bkfiles)
